ms/topicManagementService/src/main/java/com/elysium/reddot/ms/topic/infrastructure/inbound/rest/processor/CreateTopicProcessor.java
package com.elysium.reddot.ms.topic.infrastructure.inbound.rest.processor;

import com.elysium.reddot.ms.topic.application.data.dto.ApiResponseDTO;
import com.elysium.reddot.ms.topic.application.data.dto.TopicDTO;
import com.elysium.reddot.ms.topic.application.service.TopicApplicationServiceImpl;
import lombok.AllArgsConstructor;
import org.apache.camel.Exchange;
import org.apache.camel.Processor;
import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Component;

@Component
@AllArgsConstructor
public class CreateTopicProcessor implements Processor {

    private final TopicApplicationServiceImpl topicService;

    @Override
    public void process(Exchange exchange) {
        TopicDTO topicDTO = exchange.getIn().getBody(TopicDTO.class);
        TopicDTO topicCreated = topicService.createTopic(topicDTO);
        ApiResponseDTO apiResponseDTO = new ApiResponseDTO(HttpStatus.CREATED.value(),
                "Topic with name " + topicCreated.getName() + " created successfully", topicCreated);
        exchange.getMessage().setBody(apiResponseDTO);
    }
}