ms/topicManagementService/src/test/java/com/elysium/reddot/ms/topic/infrastructure/inbound/rest/processor/CreateTopicProcessorTest.java
package com.elysium.reddot.ms.topic.infrastructure.inbound.rest.processor;

import com.elysium.reddot.ms.topic.application.data.dto.ApiResponseDTO;
import com.elysium.reddot.ms.topic.application.data.dto.TopicDTO;
import com.elysium.reddot.ms.topic.application.service.TopicApplicationServiceImpl;
import org.apache.camel.CamelContext;
import org.apache.camel.Exchange;
import org.apache.camel.impl.DefaultCamelContext;
import org.apache.camel.support.DefaultExchange;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class CreateTopicProcessorTest {

    private CreateTopicProcessor createTopicProcessor;

    @Mock
    private TopicApplicationServiceImpl topicService;

    private CamelContext camelContext;

    @BeforeEach
    void setUp() {
        camelContext = new DefaultCamelContext();
        createTopicProcessor = new CreateTopicProcessor(topicService);
    }

    @Test
    @DisplayName("given valid topic when createTopic is called then topic created successfully")
    void givenValidTopic_whenCreateTopic_thenTopicCreatedSuccessfully() {
        // arrange
        TopicDTO topicDTO = new TopicDTO(null, "name", "Name", "Topic description");
        TopicDTO createdTopicDTO = new TopicDTO(1L, "name", "Name", "Topic description");

        ApiResponseDTO expectedApiResponse = new ApiResponseDTO(HttpStatus.CREATED.value(),
                "Topic with name " + createdTopicDTO.getName() + " created successfully", createdTopicDTO);

        Exchange exchange = new DefaultExchange(camelContext);
        exchange.getIn().setHeader("CamelHttpUri", "/topics");
        exchange.getIn().setBody(topicDTO);

        // mock
        when(topicService.createTopic(topicDTO)).thenReturn(createdTopicDTO);

        // act
        createTopicProcessor.process(exchange);
        ApiResponseDTO actualApiResponse = exchange.getMessage().getBody(ApiResponseDTO.class);

        // assert
        assertEquals(expectedApiResponse.getStatus(), actualApiResponse.getStatus());
        assertEquals(expectedApiResponse.getMessage(), actualApiResponse.getMessage());
        assertEquals(expectedApiResponse.getData(), actualApiResponse.getData());
    }


}
